backend/app/utils/image.py
from PIL import Image
import io
import numpy as np
from tensorflow.keras.preprocessing import image

def preprocess_image(file_content, target_size=(224, 224)):
    """
    Reads bytes, converts to image, resizes, and normalizes for model input.
    """
    img = Image.open(io.BytesIO(file_content))
    img = img.resize(target_size)
    img_array = image.img_to_array(img)
    img_array = np.expand_dims(img_array, axis=0)
    img_array /= 255.0
    return img_array
